requirements.txt
faster-whisper
grpcio
grpcio-tools
numpy
soundfile
librosa
pyyaml
sounddevice
fastapi
uvicorn
mypy-protobuf
